Move middleware code to middleware/index
Refs #341
